feat: world-view-contract-guard-meta-test-retirement-004 - Verify focused backend test evidence
